minor changes in media section

+++ Controllers/MediaPostController.js
@@ -3,10 +3,11 @@ const MediaPost = require("../Models/MediaPost");
 //Create New Media Post
 
 exports.GetPostById = async (req, res) => {
+
   const postId = req.params.id;
 
   try {
-    const post = await MediaPost.findById(postId);
+    const post = await MediaPost.findById(postId).populate("author", "ngo_name");
 
     if (!post) {
       return res
@@ -14,7 +15,11 @@ exports.GetPostById = async (req, res) => {
         .json({ success: false, message: "Post not found !!" });
     }
 
-    res.status(200).json({ success: true, postData: post });
+    const postData = { ...post._doc, author: post.author.ngo_name };
+
+    // console.log(postData);
+
+    res.status(200).json({ success: true, postData });
   } catch (err) {
     console.error(err);
     return res
@@ -24,12 +29,11 @@ exports.GetPostById = async (req, res) => {
 };
 
 exports.GetPosts = async (req, res) => {
-  const authorId = req.user._id;
-
-  console.log(req.userType);
 
   try {
-    const posts = await MediaPost.find({ author: authorId });
+    const posts = req.userType !== "ngo"
+      ? await MediaPost.find().populate("author", "ngo_name")
+      : await MediaPost.find({ author: req.user._id }).populate("author", "ngo_name");
 
     if (!posts) {
       return res
@@ -37,12 +41,18 @@ exports.GetPosts = async (req, res) => {
         .json({ success: false, message: "No posts found !!" });
     }
 
-    return res.status(200).json({ success: true, postsData: posts });
+    const postsData = posts.map((post) => {
+      const postData = { ...post._doc, author: post.author.ngo_name };
+      return postData;
+    });
+
+    return res.status(200).json({ success: true, postsData });
   } catch (err) {
     console.error(err);
     res.status(500).json({ success: false, message: "Internal server error." });
   }
 };
+
 
 exports.CreatePost = async (req, res) => {
   try {
@@ -53,9 +63,16 @@ exports.CreatePost = async (req, res) => {
         .send({ success: false, message: "Not Authorized." });
     }
 
+
     const user = req.user;
 
     const { title, content } = req.body;
+
+    if (!title || !content) {
+      return res
+        .status(400)
+        .json({ success: false, message: "Title and content is require to create a post !!!" });
+    }
 
     const newMediaPost = new MediaPost({
       title,
@@ -76,10 +93,23 @@ exports.CreatePost = async (req, res) => {
 
 exports.UpdatePost = async (req, res) => {
   try {
+
+    if (req.userType !== "ngo") {
+      return res
+        .status(401)
+        .send({ success: false, message: "Not Authorized." });
+    }
+
     const postId = req.params.id;
     const { title, content, mediaUrl } = req.body;
 
-    const media = await MediaPost.findById(postId);
+    if (!title || !content) {
+      return res
+        .status(400)
+        .json({ success: false, message: "Title and content is require to create a post !!!" });
+    }
+
+    const media = await MediaPost.findById(postId).populate("author", "ngo_name");
 
     if (!media) {
       return res
@@ -103,7 +133,11 @@ exports.UpdatePost = async (req, res) => {
 
     const updatedMedia = await media.save();
 
-    return res.status(200).json(updatedMedia);
+    const postData = { ...updatedMedia._doc, author: updatedMedia.author.ngo_name };
+
+    // console.log(postData);
+
+    return res.status(200).json({ success: true, postData });
   } catch (err) {
     return res
       .status(500)
@@ -113,6 +147,12 @@ exports.UpdatePost = async (req, res) => {
 
 exports.DeletePost = async (req, res) => {
   try {
+    if (req.userType !== "ngo") {
+      return res
+        .status(401)
+        .send({ success: false, message: "Not Authorized." });
+    }
+
     const postId = req.params.id;
 
     const deletedMedia = await MediaPost.findByIdAndDelete(postId);
@@ -123,7 +163,7 @@ exports.DeletePost = async (req, res) => {
         .json({ success: false, message: "Media post not found" });
     }
 
-    return res.status(200).json({ message: "Media post deleted successfully" });
+    return res.status(200).json({ success: true, message: "Media post deleted successfully" });
   } catch (err) {
     console.log(err);
     return res
@@ -133,6 +173,12 @@ exports.DeletePost = async (req, res) => {
 };
 
 exports.uploadFile = (req, res) => {
+
+  if (req.userType !== "ngo") {
+    return res
+      .status(401)
+      .send({ success: false, message: "Not Authorized." });
+  }
 
   if (!req.fileUrl) {
     return res.status(400).json({ success: false, message: 'No file uploaded in route' });
